move tests to enterprise
