+++ tests/local_testing/test_router_budget_limiter.py
@@ -237,80 +237,6 @@ async def test_get_budget_config_for_provider():
 
 
 @pytest.mark.asyncio
-async def test_prometheus_metric_tracking():
-    """
-    Test that the Prometheus metric for provider budget is tracked correctly
-    """
-    cleanup_redis()
-    try:
-        from litellm_enterprise.integrations.prometheus import PrometheusLogger
-        from unittest.mock import MagicMock
-    except Exception:
-        PrometheusLogger = None
-    if PrometheusLogger is None:
-        pytest.skip("PrometheusLogger is not installed")
-
-    # Create a mock PrometheusLogger
-    mock_prometheus = MagicMock(spec=PrometheusLogger)
-
-    # Setup provider budget limiting
-    provider_budget = RouterBudgetLimiting(
-        dual_cache=DualCache(),
-        provider_budget_config={
-            "openai": BudgetConfig(budget_duration="1d", max_budget=100)
-        },
-    )
-
-    litellm._async_success_callback = [mock_prometheus]
-
-    provider_budget_config: GenericBudgetConfigType = {
-        "openai": BudgetConfig(budget_duration="1d", max_budget=0.000000000001),
-        "azure": BudgetConfig(budget_duration="1d", max_budget=100),
-    }
-
-    router = Router(
-        model_list=[
-            {
-                "model_name": "gpt-3.5-turbo",  # openai model name
-                "litellm_params": {  # params for litellm completion/embedding call
-                    "model": "azure/chatgpt-v-3",
-                    "api_key": os.getenv("AZURE_API_KEY"),
-                    "api_version": os.getenv("AZURE_API_VERSION"),
-                    "api_base": os.getenv("AZURE_API_BASE"),
-                },
-                "model_info": {"id": "azure-model-id"},
-            },
-            {
-                "model_name": "gpt-3.5-turbo",  # openai model name
-                "litellm_params": {
-                    "model": "openai/gpt-4o-mini",
-                },
-                "model_info": {"id": "openai-model-id"},
-            },
-        ],
-        provider_budget_config=provider_budget_config,
-        redis_host=os.getenv("REDIS_HOST"),
-        redis_port=int(os.getenv("REDIS_PORT")),
-        redis_password=os.getenv("REDIS_PASSWORD"),
-    )
-
-    try:
-        response = await router.acompletion(
-            messages=[{"role": "user", "content": "Hello, how are you?"}],
-            model="openai/gpt-4o-mini",
-            mock_response="hi",
-        )
-        print(response)
-    except Exception as e:
-        print("error", e)
-
-    await asyncio.sleep(2.5)
-
-    # Verify the mock was called correctly
-    mock_prometheus.track_provider_remaining_budget.assert_called()
-
-
-@pytest.mark.asyncio
 async def test_handle_new_budget_window():
     """
     Test _handle_new_budget_window helper method
